+++ NEWS.md
@@ -1,3 +1,14 @@
+# SEAPSUTWorkflow 0.1.2 (2020-07-24)
+
+* New tests for new features and old bugs.
+    * Now up to 134 tests, all passing.
+    * Test coverage remains at 100%.
+* `drake` workflow now extends to completed eta_fu tables.
+* New function `assemble_eta_fu_tables()`, 
+  a counterpart to `assemble_fu_allocation_tables()`, 
+  which allows the use of exemplar countries to complete efficiency tables.
+
+
 # SEAPSUTWorkflow 0.1.1 (2020-07-24)
 
 * Workflows can now be created with additional template countries that